Move tagger_instance directly to picard/__init__.py
