+++ picard/acoustid/manager.py
@@ -26,12 +26,12 @@
 
 from functools import partial
 
-from picard import log
-from picard.i18n import N_
-from picard.util import (
-    load_json,
+from picard import (
+    log,
     tagger_instance,
 )
+from picard.i18n import N_
+from picard.util import load_json
 
 from picard.ui.enums import MainAction
