+++ picard/browser/server.py
@@ -45,15 +45,13 @@ from picard import (
     PICARD_ORG_NAME,
     PICARD_VERSION_STR,
     log,
+    tagger_instance,
 )
 from picard.browser import addrelease
 from picard.config import get_config
 from picard.const import BROWSER_INTEGRATION_LOCALIP
 from picard.oauth import OAuthInvalidStateError
-from picard.util import (
-    mbid_validate,
-    tagger_instance,
-)
+from picard.util import mbid_validate
 from picard.util.thread import to_main
