+++ picard/ui/options/dialog.py
@@ -38,7 +38,10 @@ from PyQt6 import (
     QtWidgets,
 )
 
-from picard import log
+from picard import (
+    log,
+    tagger_instance,
+)
 from picard.config import (
     Option,
     OptionError,
@@ -57,7 +60,6 @@ from picard.profile import (
 from picard.util import (
     get_url,
     restore_method,
-    tagger_instance,
 )
 
 from picard.ui import (
